Added MemoryAddress as to RegisterReference in case there is only an offset

+++ absynt.x86/src/main/scala/org/werner/absynt/x86/operands/memoryaccess/RegisterMemoryLocation.scala
@@ -36,7 +36,7 @@ class RegisterMemoryLocation[Size <: WordDoubleQuadSize](
       else if (reference.displacement.toShort == reference.displacement)
         Some(ImmediateValue.wordImmediate(reference.displacement.toShort))
       else
-        Some(ImmediateValue.doubleWordImmediate(reference.displacement)),
+        Some(ImmediateValue.doubleWordImmediate(reference.displacement.toInt)),
       reference.segment,
     )
     with ModRMEncodableOperand {
